yard-management/requirements.txt
tqdm
numpy
opencv-python
Pillow
inference
python-doctr
supervision